% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW6_Class.R
\name{coerce-numeric-sparse_numeric}
\alias{coerce-numeric-sparse_numeric}
\alias{coerce,numeric,sparse_numeric-method}
\title{Coerce numeric to sparse_numeric}
\arguments{
\item{from}{A numeric vector.}
}
\value{
A sparse_numeric object.
}
\description{
Convert a dense numeric vector to a sparse_numeric vector.
}
