% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW6_Class.R
\name{coerce-sparse_numeric-numeric}
\alias{coerce-sparse_numeric-numeric}
\alias{coerce,sparse_numeric,numeric-method}
\title{Coerce sparse_numeric to numeric}
\arguments{
\item{from}{A sparse_numeric object.}
}
\value{
A numeric vector.
}
\description{
Convert a sparse_numeric vector to a dense numeric vector.
}
