% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW6_Class.R
\name{length,sparse_numeric-method}
\alias{length,sparse_numeric-method}
\title{Sparse numeric length}
\usage{
\S4method{length}{sparse_numeric}(x)
}
\arguments{
\item{x}{A sparse_numeric object.}
}
\value{
Integer length.
}
\description{
Length of a sparse_numeric object.
}
