% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW6_Class.R
\name{mean,sparse_numeric-method}
\alias{mean,sparse_numeric-method}
\title{Sparse numeric mean}
\usage{
\S4method{mean}{sparse_numeric}(x, ...)
}
\arguments{
\item{x}{A sparse_numeric object.}

\item{...}{Ignored.}
}
\value{
A numeric scalar.
}
\description{
Mean of a sparse_numeric vector, including zero entries.
}
