% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW6_Class.R
\name{norm}
\alias{norm}
\alias{norm,sparse_numeric-method}
\title{Euclidean norm}
\usage{
norm(x, ...)

\S4method{norm}{sparse_numeric}(x, ...)
}
\arguments{
\item{x}{A sparse_numeric object.}

\item{...}{Further arguments (unused).}
}
\value{
A numeric scalar giving the norm.
}
\description{
Generic for computing the Euclidean (L2) norm.

Euclidean norm of a sparse_numeric vector.
}
