% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW6_Class.R
\name{plot,sparse_numeric,sparse_numeric-method}
\alias{plot,sparse_numeric,sparse_numeric-method}
\title{sparse numeric plot}
\usage{
\S4method{plot}{sparse_numeric,sparse_numeric}(x, y)
}
\arguments{
\item{x}{A sparse_numeric object.}

\item{y}{A sparse_numeric object.}
}
\value{
A plot.
}
\description{
Creates a scatterplot showing the values of two sparse_numeric vectors
at positions where both vectors have non-zero entries
}
