% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW6_Class.R
\name{show,sparse_numeric-method}
\alias{show,sparse_numeric-method}
\title{sparse numeric display}
\usage{
\S4method{show}{sparse_numeric}(object)
}
\arguments{
\item{object}{A sparse_numeric object.}
}
\value{
Prints summary output.
}
\description{
Prints a summary of the sparse_numeric vector
}
