% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW6_Class.R
\name{sparse_add}
\alias{sparse_add}
\title{Addition for sparse_numeric vectors}
\usage{
sparse_add(x, y)
}
\arguments{
\item{x, }{sparse numeric vector}

\item{y, }{sparse numeric vector}
}
\value{
A \code{sparse_numeric} containing the elementwise sum.
}
\description{
Add two sparse vectors
}
