% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW6_Class.R
\name{sparse_crossprod}
\alias{sparse_crossprod}
\title{Cross product for sparse_numeric vectors}
\usage{
sparse_crossprod(x, y)
}
\arguments{
\item{x, }{sparse numeric vector}

\item{y, }{sparse numeric vector}
}
\value{
A numeric scalar giving the dot product.
}
\description{
Sparse crossproduct (dot product)
}
