% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW6_Class.R
\name{sparse_mult}
\alias{sparse_mult}
\title{Multiplication for sparse_numeric vectors}
\usage{
sparse_mult(x, y)
}
\arguments{
\item{x, }{sparse numeric vector}

\item{y, }{sparse numeric vector.}
}
\value{
A \code{sparse_numeric} containing the elementwise product.
}
\description{
Elementwise multiply two sparse vectors
}
