% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW6_Class.R
\docType{class}
\name{sparse_numeric-class}
\alias{sparse_numeric-class}
\title{Sparse numeric vector class}
\description{
An S4 class for representing numeric vectors in sparse form.
}
\section{Slots}{

\describe{
\item{\code{value}}{Numeric vector of non-zero values.}

\item{\code{pos}}{Integer vector of positions of the non-zero values.}

\item{\code{length}}{Single integer giving the length of the full vector.}
}}

