% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW6_Class.R
\name{sparse_sub,sparse_numeric,sparse_numeric-method}
\alias{sparse_sub,sparse_numeric,sparse_numeric-method}
\title{Subtraction for sparse_numeric vectors}
\usage{
\S4method{sparse_sub}{sparse_numeric,sparse_numeric}(x, y)
}
\arguments{
\item{x}{A sparse_numeric object.}

\item{y}{A sparse_numeric object.}
}
\value{
A sparse_numeric object containing the difference.
}
\description{
Subtract one sparse_numeric vector from another.
}
