% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW6_Class.R
\name{sparse_sub}
\alias{sparse_sub}
\title{Subtraction for sparse_numeric vectors}
\usage{
sparse_sub(x, y)
}
\arguments{
\item{x, }{sparse numeric vector}

\item{y, }{sparse numeric vector}
}
\value{
A \code{sparse_numeric} containing the elementwise difference.
}
\description{
Subtract two sparse vectors
}
