% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW6_Class.R
\name{standardize}
\alias{standardize}
\alias{standardize,sparse_numeric-method}
\title{Standardize sparse_numeric vector}
\usage{
standardize(x, ...)

\S4method{standardize}{sparse_numeric}(x, ...)
}
\arguments{
\item{x}{A sparse_numeric object.}

\item{...}{Ignored.}
}
\value{
A sparse_numeric object.
}
\description{
Computes (x - mean) / sd without dense coercion.

Standardize a sparse_numeric vector.
}
\details{
Computes (x - mean) / sd over all entries including zeros.
}
