% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW6_Class.R
\name{*,sparse_numeric,sparse_numeric-method}
\alias{*,sparse_numeric,sparse_numeric-method}
\title{\itemize{
\item operator
}}
\usage{
\S4method{*}{sparse_numeric,sparse_numeric}(e1, e2)
}
\arguments{
\item{e1}{A sparse_numeric object.}

\item{e2}{A sparse_numeric object.}
}
\value{
A sparse_numeric object.
}
\description{
Multiply two sparse_numeric objects using the * operator.
}
